process_upload/requirements.txt
google-cloud-storage==2.11.0
google-cloud-firestore==2.11.0
pdfminer.six==20221105
python-docx==0.8.11
